♻️ refactor(db): Change db/ directory hierarchy
