+++ src/db/methods/collections/__init__.py
@@ -1,0 +1,20 @@
+"""This module for working with collections"""
+
+from . collections import users, \
+                            roles, \
+                            domains, \
+                            contests, \
+                            group_roles, \
+                            group_members, \
+                            internal_counters
+
+
+__all__ = [
+    "users",
+    "roles",
+    "domains",
+    "contests",
+    "group_roles",
+    "group_members",
+    "internal_counters"
+]
