+++ src/db/methods/user.py
@@ -1,9 +1,13 @@
 """Methods for interact with users in database"""
-from db import db
 from loguru import logger
 
+from db.client import db
+from . import collections
+
+
 def test():
-    """Test method function"""
-    logger.info("test")
+    """Test method"""
+
     logger.info(db)
-    
+    logger.info(collections.users)
+    logger.info(collections.internal_counters)
